Add AE Dashboard API and fix DoseCoverage cast issue
Expanded internal documentation for AE dashboard data-access files.
Implemented `AdverseEventController` with endpoints for product catalog,
favorites, triage, forest, quadrant, reverse lookup, and interchange.
Enforced feature flags, input validation, and authentication.

Fixed `DoseCoverage` cast issue in `LabelView.cs` by aligning EF materialization
with SQL Server's `decimal` type and converting to `double` in DTOs.

Added comprehensive tests in `AdverseEventControllerTests.cs` for API coverage,
including feature gating, metadata, and Swagger validation.

Introduced helper methods for feature flag checks, user resolution, and
favorite state mutations.

+++ MedRecPro/DataAccess/DtoLabelAccess-AeDashboard.cs
@@ -777,7 +777,7 @@ namespace MedRecPro.DataAccess
                 // picker filtering.
                 PlaceboCoverage = entity.PlaceboCoverage,
                 ActiveCoverage = entity.ActiveCoverage,
-                DoseCoverage = entity.DoseCoverage,
+                DoseCoverage = (double)entity.DoseCoverage,
                 SocBreadth = entity.SocBreadth,
 
                 // Fall back to dashboard metadata when the view does not provide a
